requirements.txt
streamlit==1.43.2
langchain==0.3.20
faiss-cpu ==1.10.0
langchain-core==0.3.45
langchain-google-genai ==2.1.0
PyPDF2==3.0.1
google-ai-generativelanguage==0.6.17
langchain-community==0.3.19
pandas==2.2.3


wikipedia
sentence-transformers
transformers